+++ src/util/const.ts
@@ -1,7 +1,6 @@
 import { Network } from '../types/types';
 import { Network as NetworkFromEthers } from '@ethersproject/networks';
 
-export const DEFAULT_CONTRACT_ADDRESSES = 'DEFAULT_TOKENS';
 export const DEFAULT_ALCHEMY_API_KEY = 'demo';
 export const DEFAULT_NETWORK = Network.ETH_MAINNET;
 export const DEFAULT_MAX_RETRIES = 5;