+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Major Changes
 - POTENTIALLY BREAKING: Fixed a typing bug where the `totalSupply` field in an `NftContract` should have type `string` instead of `number`.
+- Added commonly used utility methods from ethers.js into a top-level `Utils` export.  
 
 ## 2.0.4